requirements.txt
# Flask core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Environment & Config
python-dotenv==1.0.0

# Password hashing
bcrypt==4.1.1

# Email
sendgrid==6.11.0

# Scheduling
APScheduler==3.10.4

# Image processing
Pillow==10.1.0

# Google Places (for later)
googlemaps==4.10.0

# Maps & Charts (for later)
folium==0.15.1
plotly==5.18.0

# Production server
gunicorn==21.2.0
